+++ CHANGELOG.md
@@ -4,5 +4,6 @@ Changelog
 [Unreleased]
 ------------
 
--  Add apikey model (#9)
--  Add Django admin to manage API keys (#11)
+- Add apikey model (#9)
+- Add Django admin to manage API keys (#11)
+- Add authentication backend (#12)